added playlist functionality

+++ lib/files/downloads.dart
@@ -8,22 +8,35 @@ import 'package:path_provider/path_provider.dart';
 import 'package:voices_for_christ/database/local_db.dart';
 import 'package:voices_for_christ/data_models/message_class.dart';
 
-import 'package:fluttertoast/fluttertoast.dart';
-import 'package:flutter/material.dart';
+//import 'package:fluttertoast/fluttertoast.dart';
+//import 'package:flutter/material.dart';
 
-Future<void> deleteMessageFile(Message message) async {
+Future<Message> deleteMessageFile(Message message) async {
   String dir = '';
   String filepath = '';
+  final db = MessageDB.instance;
 
   try {
+    print('DELETE: starting');
     dir = (await getApplicationDocumentsDirectory()).path;
     filepath = '$dir/${message.id.toString()}.mp3';
-  } catch (error) {
-    print(error);
-  }
+    File f = File('$filepath');
+    if (await f.exists()) {
+      f.delete();
+    }
+    print('DELETE: deleted file');
 
-  File f = File('$filepath');
-  f.delete();
+    print('DELETE: updating message info');
+    message.isdownloaded = 0;
+    message.filepath = '';
+    await db.update(message);
+    print('DELETE: message info updated: ${message.isdownloaded} and ${message.filepath}');
+
+    return message;
+  } catch (error) {
+    print('Error deleting file: $error');
+    throw Exception(error);
+  }
 }
 
 Future<Message> downloadMessageFile(Message message) async {
@@ -59,7 +72,7 @@ Future<Message> downloadMessageFile(Message message) async {
     message.iscurrentlydownloading = 0;
     await db.update(message);
 
-    Fluttertoast.showToast(
+    /*Fluttertoast.showToast(
       msg: "Error downloading: check connection",
       toastLength: Toast.LENGTH_LONG,
       gravity: ToastGravity.BOTTOM,
@@ -67,7 +80,7 @@ Future<Message> downloadMessageFile(Message message) async {
       backgroundColor: Colors.red[600],
       textColor: Colors.grey[50],
       fontSize: 16.0
-    );
+    );*/
 
     throw Exception(error);
   }
